Add deployment of Azure Bakery (Kubernetes based) (#6)
TL;DR: This PR adds the capability to deploy a complete Pangeo Forge Azure Bakery (k8s based) from a clean slate, all you need is an Azure Account (with what I believe to be pretty wide stretching permissions 😅 ). This PR creates all the Terraform resources & scripts needed to deploy the base infrastructure to support AKS and then the application of the Prefect K8s configuration on top. A Pangeo Forge Recipe that is setup as a flow is also provided to register & run against the Bakery to exercise it.